// src/Connections/type.ts
/**
 * Overall category for a connection. Used to enable / disable
 * specific connection classes and organise commands.
 */
export enum ConnectionType {
  ChallengeHound = "challengehound",
  Feeds = "feeds",
  Figma = "figma",
  Generic = "generic",
  GenericOutbound = "genericOutbound",
  Github = "github",
  Gitlab = "gitlab",
  Jira = "jira",
  OpenProject = "openproject",
}